Fix mobile bottom bar and show joined whops in sidebar
