+++ src/components/BottomBar.jsx
@@ -6,21 +6,20 @@ import '../styles/bottombar.scss';
 import { FiMenu, FiSun, FiMoon } from 'react-icons/fi';
 import { FaSignOutAlt } from 'react-icons/fa';
 import { useTheme } from '../context/ThemeContext';
+import useJoinedWhops from '../hooks/useJoinedWhops';
 
 export default function BottomBar() {
   const { theme, setLight, setDark } = useTheme();
   const { logout } = useAuth();
   const [dropdownOpen, setDropdownOpen] = useState(false);
   const [hoveredX, setHoveredX] = useState(null);
-  const [joinedWhops, setJoinedWhops] = useState([]);
-  const [loadingWhops, setLoadingWhops] = useState(true);
+  const { joinedWhops, loadingWhops } = useJoinedWhops();
   const [balance, setBalance] = useState(0);
 
   const iconsContainerRef = useRef(null);
   const navigate = useNavigate();
 
   useEffect(() => {
-    fetchJoinedWhops();
     fetchProfileBalance();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
@@ -48,67 +47,6 @@ export default function BottomBar() {
   const formatBalance = amount =>
     amount >= 1000 ? `${(amount / 1000).toFixed(2)}k` : amount.toFixed(2);
 
-  // Load Whops where the user is a member or owner
-  const fetchJoinedWhops = async () => {
-    setLoadingWhops(true);
-    try {
-      // 1) Whops where the user is a member
-      const resMembers = await fetch('https://app.byxbot.com/php/get_joined_whops.php', {
-        method: 'GET',
-        credentials: 'include',
-      });
-      if (!resMembers.ok) throw new Error(`Members HTTP ${resMembers.status}`);
-      const membersJson = await resMembers.json();
-      // pokud API vrací { status, data: [...] }
-      let membersData = Array.isArray(membersJson)
-        ? membersJson
-        : Array.isArray(membersJson.data)
-        ? membersJson.data
-        : [];
-      
-      // 2) Whops where the user is the owner
-      const resOwned = await fetch('https://app.byxbot.com/php/get_whop.php?owner=me', {
-        method: 'GET',
-        credentials: 'include',
-      });
-      if (!resOwned.ok) throw new Error(`Owned HTTP ${resOwned.status}`);
-      const ownedJson = await resOwned.json();
-      if (ownedJson.status !== 'success' || !Array.isArray(ownedJson.data)) {
-        throw new Error('Failed to load owned Whops');
-      }
-      const ownedData = ownedJson.data;
-
-      // 3) Sloučení obou seznamů bez duplicit (klíč slug)
-      const mapBySlug = new Map();
-      membersData.forEach(w => {
-        if (w?.slug) {
-          mapBySlug.set(w.slug, {
-            id: w.whop_id ?? w.id,
-            slug: w.slug,
-            banner_url: w.banner_url,
-            name: w.name ?? w.slug,
-          });
-        }
-      });
-      ownedData.forEach(w => {
-        if (w?.slug) {
-          mapBySlug.set(w.slug, {
-            id: w.id,
-            slug: w.slug,
-            banner_url: w.banner_url,
-            name: w.name ?? w.slug,
-          });
-        }
-      });
-
-      setJoinedWhops(Array.from(mapBySlug.values()));
-    } catch (err) {
-      console.error('Error loading joined/owned Whops:', err);
-      setJoinedWhops([]);
-    } finally {
-      setLoadingWhops(false);
-    }
-  };
 
   const handleMenuClick = () => {
     setDropdownOpen(prev => !prev);
