+++ src/styles/bottombar.scss
@@ -20,15 +20,7 @@
   justify-content: space-between;
   box-shadow: var(--shadow-sm);
 
-  @include respond(sm) {
-    flex-direction: column;
-    width: var(--sidebar-width);
-    height: 100vh;
-    top: 0;
-    bottom: 0 !important;
-    padding: var(--spacing-md) var(--spacing-sm);
-    justify-content: flex-start;
-  }
+
 
   &__left {
     position: relative;
@@ -218,6 +210,7 @@
     gap: var(--spacing-md);
     overflow: visible;
     z-index: 9999;
+    @include respond(sm) { display: none; }
   }
 
   &__center-icon {
